+++ packages/civ7-direct-control/src/index.ts
@@ -14,6 +14,14 @@ import {
   parseCiv7TunerFrame,
   type Civ7TunerFrame,
 } from "./session/framing.js";
+import type {
+  Civ7CommandResult,
+  Civ7DirectControlEndpoint,
+  Civ7DirectControlOptions,
+  Civ7TunerState,
+  Civ7TunerStateRole,
+  Civ7TunerStateSelection,
+} from "./session/types.js";
 import {
   CIV7_TUNER_APP_UI_STATE_NAME,
   CIV7_TUNER_STATE_NAME,
@@ -174,6 +182,14 @@ export {
   parseCiv7TunerFrame,
 } from "./session/framing.js";
 export type { Civ7TunerFrame } from "./session/framing.js";
+export type {
+  Civ7CommandResult,
+  Civ7DirectControlEndpoint,
+  Civ7DirectControlOptions,
+  Civ7TunerState,
+  Civ7TunerStateRole,
+  Civ7TunerStateSelection,
+} from "./session/types.js";
 export {
   CIV7_TUNER_APP_UI_STATE_NAME,
   CIV7_TUNER_STATE_NAME,
@@ -254,42 +270,7 @@ export const DEFAULT_CIV7_SINGLE_PLAYER_SAVE_DIR = join(
   "Single",
 );
 
-export type Civ7TunerState = Readonly<{
-  id: string;
-  name: string;
-}>;
-
-export type Civ7TunerStateRole = "app-ui" | "tuner";
-
-export type Civ7TunerStateSelection =
-  | string
-  | Readonly<{
-      id?: string;
-      name?: string;
-      role?: Civ7TunerStateRole;
-    }>;
-
-export type Civ7DirectControlEndpoint = Readonly<{
-  host: string;
-  port: number;
-}>;
-
-export type Civ7DirectControlOptions = Readonly<{
-  host?: string;
-  hosts?: ReadonlyArray<string>;
-  port?: number;
-  timeoutMs?: number;
-  env?: NodeJS.ProcessEnv;
-}>;
-
 export type Civ7UiLoadingStateName = keyof typeof CIV7_UI_LOADING_STATES;
-
-export type Civ7CommandResult = Readonly<{
-  host: string;
-  port: number;
-  state: Civ7TunerState;
-  output: ReadonlyArray<string>;
-}>;
 
 export type Civ7RuntimeApiRoot = Readonly<{
   name: string;
